split name config and utils
